[cudamapper] Renamed application_parameters.cu -> application_parameters.cpp and application_parameters.cuh -> application_parameters.hpp
