+++ cudamapper/CMakeLists.txt
@@ -32,7 +32,7 @@ else()
 endif()
 
 cuda_add_library(cudamapper
-        src/application_parameters.cu
+        src/application_parameters.cpp
         src/cudamapper.cpp
         src/index_batcher.cu
         src/index_descriptor.cpp
