+++ cudamapper/src/application_parameters.hpp
